requirements.txt
numpy>=1.21.0
pandas>=1.3.0
requests>=2.26.0
pytest>=6.2.5
mysql-connector-python>=8.0.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy-utils>=0.38.0
aiohttp>=3.8.0
ray[default]>=2.0.0
ray[tune]>=2.0.0
scikit-learn>=1.0.0
fastmcp>=2.11.0
psutil>=5.9.0
langgraph
langchain[anthropic,deepseek]
python-dotenv>=0.19.0
openai>=1.0
IPython
langsmith
langgraph-checkpoint-sqlite
langgraph-supervisor
agno
matplotlib
langchain_openai
langchain_core
pydantic
mcp
sqlalchemy
fastapi[standard]
langchain-mcp-adapters